+++ .envrc
@@ -1,0 +1,4 @@
+export FLASK_APP=dash_app
+export FLASK_ENV=development
+export DATABASE_URL=sqlite:///$PWD/app.db
+export PORTDASH_CONF=sample_config.yaml
