+++ portdash/extensions.py
@@ -1,0 +1,5 @@
+from flask_migrate import Migrate
+from flask_sqlalchemy import SQLAlchemy
+
+db = SQLAlchemy()
+migrate = Migrate()
