+++ portdash/models/__init__.py
@@ -1,0 +1,1 @@
+from app.extensions import db
