+++ portdash/models/base.py
@@ -1,0 +1,4 @@
+from flask_sqlalchemy import SQLAlchemy
+
+
+db = SQLAlchemy()
