+++ portdash/webapp.py
@@ -1,0 +1,8 @@
+from flask import Blueprint
+
+server_bp = Blueprint('main', __name__)
+
+
+@server_bp.route('/')
+def index():
+    return
