+++ requirements.txt
@@ -1,0 +1,8 @@
+pyyaml
+pandas
+numpy
+dash
+requests
+flask
+flask-sqlalchemy
+flask-migrate
